requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
api==0.0.7
argcomplete==3.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-search-documents==11.4.0
azure-storage-blob==12.19.1
Babel==2.13.1
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorclass==2.2.2
comm==0.2.0
cryptography==42.0.5
dataclasses-json==0.6.4
debugpy==1.8.0
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
emoji==2.10.1
enum34==1.1.10
executing==2.0.1
fastjsonschema==2.19.0
filetype==1.2.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
fqdn==1.5.1
frozenlist==1.4.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
inflection==0.5.1
ipykernel==6.27.1
ipython==8.18.1
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.12.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kaggle==1.6.6
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.23
loguru==0.7.2
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
msal==1.27.0
msal-extensions==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
nose==1.3.7
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
openai==1.12.0
openapi==1.1.0
ordered-set==4.1.0
orjson==3.9.15
overrides==7.4.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.2.0
pinecone-client==3.1.0
pip-upgrader==1.4.15
platformdirs==4.1.0
portalocker==2.8.2
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.0.9
pypdf==4.1.0
PyPDF2==3.0.1
pypdfium2==4.28.0
PyPika==0.48.9
python-dateutil==2.8.2
    python-dotenv==1.0.1
python-iso639==2024.2.7
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.4
pytz==2024.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.6.2
referencing==0.31.1
regex==2023.12.25
requests==2.31.0
retriever==3.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.13.2
Send2Trash==1.8.2
setuptools==69.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.20.0
streamlit==1.32.0
streamlit-chat==0.1.1
streamlit-option-menu==0.3.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
terminaltables==3.1.10
text-unidecode==1.3
tika==2.6.0
tiktoken==0.6.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
unstructured==0.11.8
unstructured-client==0.21.1
uri-template==1.3.0
urllib3==2.1.0
watchdog==4.0.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.1.1
win32-setctime==1.1.0
wrapt==1.16.0
xlrd==2.0.1
yarl==1.9.4
